% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_settings.R
\name{set_cache_length}
\alias{set_cache_length}
\alias{get_cache_length}
\title{Set and get cache validity time}
\usage{
set_cache_length(length)

get_cache_length()
}
\arguments{
\item{length}{\code{numeric} time in seconds how long the cache is valid.}
}
\value{
\code{get_cache_length()} \code{numeric} length of cache validity (default value is 7 days)
}
\description{
Set and get how long the cache is valid and the files are downloaded only once and then returned from the cache.
}
\section{Functions}{
\itemize{
\item \code{set_cache_length}: Set cache length

\item \code{get_cache_length}: Get cache length
}}

