% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Data50.R
\name{generate_Data50_citation}
\alias{generate_Data50_citation}
\title{Generate attribution for dataset Data50}
\usage{
generate_Data50_citation()
}
\value{
\code{character} with citation.
}
\description{
Create citation string as per terms of use (\url{https://geoportal.cuzk.cz/Dokumenty/Podminky_EN.pdf}).
}
\examples{
generate_Data50_citation()

}
