% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_average_salary.R
\name{load_average_salary}
\alias{load_average_salary}
\alias{load_average_salary_col_explanations}
\title{Load average salary for specific spatial units}
\usage{
load_average_salary()

load_average_salary_col_explanations()
}
\value{
data.frame containg the requested data
}
\description{
Load average salary for specific spatial units for years 2011 to 2017.
}
\section{Functions}{
\itemize{
\item \code{load_average_salary_col_explanations}: Load description for columns
}}

\examples{
\dontrun{
    average_salary <- load_average_salary()
}
}
