% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cadastre.R
\name{load_cadastral_territory}
\alias{load_cadastral_territory}
\title{Extract data from Cadastral map}
\usage{
load_cadastral_territory(id, layer = "katastralni uzemi", WGS84 = FALSE)
}
\arguments{
\item{id}{id of cadastral territory as character}

\item{layer}{identification of data to extract as character, see details.
Default value is \code{"katastralni území"}}

\item{WGS84}{convert data to WGS-84 coordinate system? Default \code{FALSE}.}
}
\value{
\code{data.frame} with spatail objects (\code{\link[sf]{sf}}) of the specified layer
}
\description{
Extract specific layer, in form of spatial data, from cadastral map for given cadastral
territory in Czech Republic. Checks are performed to find out if the provided \code{id} is valid
for some cadastral territory in Czech Republic.
}
\details{
The \code{layer} can have values from following set, the value in brackets is alias to full layer
name:
\enumerate{
\item \code{"BODOVE_POLE_B"}
\item \code{"BODOVE_POLE_T"}
\item \code{"BUDOVY_B"} (\code{"budovy body"})
\item \code{"BUDOVY_DEF"}
\item \code{"BUDOVY_P"} (\code{"budovy"})
\item \code{"DALSI_PRVKY_MAPY_B"}
\item \code{"DALSI_PRVKY_MAPY_L"}
\item \code{"DALSI_PRVKY_MAPY_T"}
\item \code{"HRANICE_PARCEL_L"} (\code{"hranice parcel"})
\item \code{"KATASTRALNI_UZEMI_DEF"}
\item \code{"KATASTRALNI_UZEMI_L"}
\item \code{"KATASTRALNI_UZEMI_P"} (\code{"katastralni uzemi"})
\item \code{"PARCELY_KN_B"}
\item \code{"PARCELY_KN_DEF"}
\item \code{"PARCELY_KN_L"}
\item \code{"PARCELY_KN_P"} (\code{"parcely"})
\item \code{"PRVKY_ORIENT_MAPY_B"}
\item \code{"PRVKY_ORIENT_MAPY_L"}
\item \code{"PRVKY_ORIENT_MAPY_T"}
\item \code{"VB_P"}
}
So the codes \code{layer = "BUDOVY_B"} and \code{layer = "budovy body"} are equal.

The values of \code{id} follow general pattern of six number with first number being 6,7 or 9.
}
\section{Information about dataset}{

More detailed information about data can be found at the provider's website
\url{http://atom.cuzk.cz/}.
}

\examples{
\dontrun{
    parcely_vyskov <- load_cadastral_territory("788571", layer = "parcely")
}
}
