% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Data200.R
\name{load_Data200}
\alias{load_Data200}
\alias{save_Data200}
\title{Load or save from Data200}
\usage{
load_Data200(layer, WGS84 = FALSE)

save_Data200(path, layer = NULL, type = NULL)
}
\arguments{
\item{layer}{identification of data to extract as character, see details.}

\item{WGS84}{convert data to WGS-84 coordinate system? Default \code{FALSE}.}

\item{path}{\code{character} path to store the files to.}

\item{type}{\code{character} type of layers to save. See details, types are listed in brackets.}
}
\value{
\code{"load_Data200"} - \code{data.frame} with spatial objects (\code{\link[sf]{sf}}) of the specified layer.
For \code{layer} either \code{"DMR"} or \code{"DMRShaded"} the output is actually a (\code{\link[raster]{raster}}).
\code{"save_Data200"} - \code{path} to the unzipped files (for layer) or folder (for type),
the zipped file is also stored at path (mainly for further use)
}
\description{
Load data from Data200 data source (\url{https://geoportal.cuzk.cz/(S(ijginumejzilvacbfijkylwj))/Default.aspx?mode=TextMeta&side=mapy_data200&text=dSady_mapyData200&head_tab=sekce-02-gp&menu=229}).
The data can be used only after correctly citing the creator (as per terms of use \url{https://geoportal.cuzk.cz/Dokumenty/Podminky_EN.pdf}).
The citation is in form "Mapový podklad – Data200, {{insert year}} © Český úřad zeměměřický a katastrální, www.cuzk.cz".
}
\details{
The \code{layer} can have values from following set, in the bracket is the name of general category
(can be used as type in saving the data):
\enumerate{
\item \code{"AdministrativniHraniceLinie"} (\code{"Hranice"})
\item \code{"AdministrativniUzemiCentroid"} (\code{"Hranice"})
\item \code{"AdministrativniUzemiUTJ"} (\code{"Hranice"})
\item \code{"AdministrativniUzemiObce"} (\code{"Hranice"})
\item \code{"AdministrativniUzemiOkresy"} (\code{"Hranice"})
\item \code{"AdministrativniUzemiKraje"} (\code{"Hranice"})
\item \code{"HrazJezNad50m"} (\code{"Vodstvo"})
\item \code{"HrazJezPod50m"} (\code{"Vodstvo"})
\item \code{"JezeroRybnikVodniNadrz"} (\code{"Vodstvo"})
\item \code{"VodniTokPod50m"} (\code{"Vodstvo"})
\item \code{"VodniTokNad50m"} (\code{"Vodstvo"})
\item \code{"Ostrovy"} (\code{"Vodstvo"})
\item \code{"MokrinaBazina"} (\code{"Vodstvo"})
\item \code{"Vodopad"} (\code{"Vodstvo"})
\item \code{"Prameny1"} (\code{"Vodstvo"})
\item \code{"Prameny2"} (\code{"Vodstvo"})
\item \code{"OrografickeNazvy"} (\code{"Popis"})
\item \code{"GeomorfologickeOblasti"} (\code{"Popis"})
\item \code{"GeomorfologickeCelky"} (\code{"Popis"})
\item \code{"GeomorfologickePodcelky"} (\code{"Popis"})
\item \code{"NarodniParkPrirodniRezervace"} (\code{"RuzneObjekty"})
\item \code{"Produktovod"} (\code{"RuzneObjekty"})
\item \code{"Vysilac"} (\code{"RuzneObjekty"})
\item \code{"VyznamneObjekty"} (\code{"RuzneObjekty"})
\item \code{"ProduktovodVyznamneBody"} (\code{"RuzneObjekty"})
\item \code{"Věž"} (\code{"RuzneObjekty"})
\item \code{"DulLom"} (\code{"RuzneObjekty"})
\item \code{"Budova"} (\code{"RuzneObjekty"})
\item \code{"ElektrickeVedeni"} (\code{"RuzneObjekty"})
\item \code{"Elektrarna"} (\code{"RuzneObjekty"})
\item \code{"ObceBody"} (\code{"Sidla"})
\item \code{"ObcePolygony"} (\code{"Sidla"})
\item \code{"Privoz"} (\code{"Doprava"})
\item \code{"PrivozStanice"} (\code{"Doprava"})
\item \code{"ZeleznicniPrejezd"} (\code{"Doprava"})
\item \code{"Heliport"} (\code{"Doprava"})
\item \code{"LanovaDraha"} (\code{"Doprava"})
\item \code{"DalnicniOdpocivka"} (\code{"Doprava"})
\item \code{"KrizovatkaMimourovnova"} (\code{"Doprava"})
\item \code{"LetisteNad40Ha"} (\code{"Doprava"})
\item \code{"LetisteNad40HaBod"} (\code{"Doprava"})
\item \code{"ZelezniceZastavky"} (\code{"Doprava"})
\item \code{"LetistePod40Ha"} (\code{"Doprava"})
\item \code{"LodniPristav"} (\code{"Doprava"})
\item \code{"PristavaciDraha"} (\code{"Doprava"})
\item \code{"Zeleznice"} (\code{"Doprava"})
\item \code{"Silnice"} (\code{"Doprava"})
\item \code{"LesyPlantaze"} (\code{"Vegetace"})
\item \code{"KotovaneBody"} (\code{"Relief"})
\item \code{"Vrstevnice"} (\code{"Relief"})
\item \code{"SkalniStenaSraz"} (\code{"Relief"})
\item \code{"Jeskyne"} (\code{"Relief"})
\item \code{"DMR"} (\code{"Relief"})
\item \code{"DMRShaded"} (\code{"Relief"})
}
}
\section{Functions}{
\itemize{
\item \code{load_Data200}: Loads single dataset

\item \code{save_Data200}: Download and store layer (and zipped general category) or complete category
}}

\examples{
\dontrun{
    waterfalls <- load_Data200(layer = "Vodopad")
}
\dontrun{
    folder_water_objects <- save_Data50("~/data/coomunications", type = "Vodopad")
}
}
