% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Data200.R, R/load_Data50.R
\name{load_Data200_info}
\alias{load_Data200_info}
\alias{load_Data50}
\alias{save_Data50}
\alias{load_Data50_info}
\title{Load or save from Data50}
\usage{
load_Data200_info(english_names = FALSE)

load_Data50(layer, WGS84 = FALSE)

save_Data50(path, layer = NULL, type = NULL)

load_Data50_info(english_names = FALSE)
}
\arguments{
\item{english_names}{change the names of the columns to English. Default \code{FALSE}.}

\item{layer}{identification of data to extract as character, see details.}

\item{WGS84}{convert data to WGS-84 coordinate system? Default \code{FALSE}.}

\item{path}{\code{character} path to store the files to.}

\item{type}{\code{character} type of layers to save. See details, types are listed in brackets.}
}
\value{
\code{"load_Data50"} - \code{data.frame} with spatial objects (\code{\link[sf]{sf}}) of the specified layer.
\code{"save_Data50"} - \code{path} to the unzipped files (for layer) or folder (for type),
the zipped file is also stored at path (mainly for further use)

\code{data.frame} with description of layers.
}
\description{
Load data from Data50 data source (\url{https://geoportal.cuzk.cz/(S(xbw0cmgh1cve4bciko2oo4e2))/Default.aspx?lng=EN&mode=TextMeta&side=mapy_data50&text=dSady_mapyData50&head_tab=sekce-02-gp&menu=2290}).
The data can be used only after correctly citing the creator (as per terms of use \url{https://geoportal.cuzk.cz/Dokumenty/Podminky_EN.pdf}).
The citation is in form "Mapový podklad – Data50, {{insert year}} © Český úřad zeměměřický a katastrální, www.cuzk.cz".

Some basic description of the dataset Data50 or Data200. Most importantly names of layers and sizes
of files that need to be downloaded.
}
\details{
The \code{layer} can have values from following set, in the bracket is the name of general category
(can be used as type in saving the data):
\enumerate{
\item \code{"BlokBudov"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"Budova"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"Hrad"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"Hrbitov"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"ChatovaKolonie"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"Kostel"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"LyzarskyMustek"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"RozhlednaVysilac"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"Rozvalina"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"Stadion"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"UsazovaciNadrzOdkaliste"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"VetrnyMotor"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"VezovitaStavba"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"VodojemVezovy"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"Zamek"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"Zricenina"} (\code{"sidelniKulturniHospodarskeObjekty"})
\item \code{"Cesta"} (\code{"komunikace"})
\item \code{"LanovaDraha"} (\code{"komunikace"})
\item \code{"Lavka"} (\code{"komunikace"})
\item \code{"Letiste"} (\code{"komunikace"})
\item \code{"LetisteObvodovaLinie"} (\code{"komunikace"})
\item \code{"Most"} (\code{"komunikace"})
\item \code{"Pesina"} (\code{"komunikace"})
\item \code{"Pristav"} (\code{"komunikace"})
\item \code{"Privoz"} (\code{"komunikace"})
\item \code{"SilniceDalnice"} (\code{"komunikace"})
\item \code{"SilniceVeVystavbe"} (\code{"komunikace"})
\item \code{"Tunel"} (\code{"komunikace"})
\item \code{"Ulice"} (\code{"komunikace"})
\item \code{"ZeleznicniStanice"} (\code{"komunikace"})
\item \code{"ZeleznicniTrat"} (\code{"komunikace"})
\item \code{"ZeleznicniVlecka"} (\code{"komunikace"})
\item \code{"ElektrickeVedeni"} (\code{"produktovodyElektrickeVedeni"})
\item \code{"Produktovod"} (\code{"produktovodyElektrickeVedeni"})
\item \code{"Akvadukt"} (\code{"vodstvo"})
\item \code{"Hraz"} (\code{"vodstvo"})
\item \code{"Jez"} (\code{"vodstvo"})
\item \code{"Shybka"} (\code{"vodstvo"})
\item \code{"VodniPlocha"} (\code{"vodstvo"})
\item \code{"VodniTok"} (\code{"vodstvo"})
\item \code{"HraniceSpravniJednotkyaKU"} (\code{"hraniceUzemnichJednotek"})
\item \code{"ChraneneUzemi"} (\code{"hraniceUzemnichJednotek"})
\item \code{"Les"} (\code{"vegetacePovrch"})
\item \code{"LoukaPastvina"} (\code{"vegetacePovrch"})
\item \code{"RaselinisteMocalBazina"} (\code{"vegetacePovrch"})
\item \code{"ZahradaSadParkViniceChmelnice"} (\code{"vegetacePovrch"})
\item \code{"Jeskyne"} (\code{"terenniRelief"})
\item \code{"KotovanyBod"} (\code{"terenniRelief"})
\item \code{"SkalnatySraz"} (\code{"terenniRelief"})
\item \code{"Skaly"} (\code{"terenniRelief"})
\item \code{"TerenniStupen"} (\code{"terenniRelief"})
\item \code{"TerenniStupenSpadnice"} (\code{"terenniRelief"})
\item \code{"Vrstevnice"} (\code{"terenniRelief"})
\item \code{"DefinicniBodCastiObce"} (\code{"popis"})
\item \code{"DefinicniBodSpravnihoCelku"} (\code{"popis"})
\item \code{"Jmeno_B"} (\code{"popis"})
\item \code{"Jmeno_L"} (\code{"popis"})
\item \code{"Jmeno_P"} (\code{"popis"})
\item \code{"ObjektRuzny"} (\code{"popis"})
}
}
\section{Functions}{
\itemize{
\item \code{load_Data200_info}: Load information about layers in Data200.

\item \code{load_Data50}: Loads single dataset

\item \code{save_Data50}: Download and store layer (and zipped general category) or complete category

\item \code{load_Data50_info}: Load information about layers in Data50.
}}

\examples{
\dontrun{
    rivers <- load_Data50(layer = "VodniTok")
}
\dontrun{
    folder_communications <- save_Data50("~/data/coomunications", type = "komunikace")
}
\dontrun{
    info <- load_Data50_info(english_names = TRUE)
    info <- load_Data200_info(english_names = TRUE)
}
}
