% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Data50.R
\name{load_Data50_info}
\alias{load_Data50_info}
\title{Loads general info about Data50 dataset}
\usage{
load_Data50_info(english_names = FALSE)
}
\arguments{
\item{english_names}{change the names of the columns to English. Default \code{FALSE}.}
}
\value{
\code{data.frame} with description of layers.
}
\description{
Some basic description of the dataset Data50. Most importantly names of layers and size of files
that need to be downloaded.
}
\examples{
\dontrun{
    info <- load_Data50_info(english_names = TRUE)
}
}
