% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_financial_indicators.R
\name{load_financial_indicators}
\alias{load_financial_indicators}
\alias{load_financial_indicators_col_explanations}
\title{Defunct: Load financial indicators for specific spatial units}
\usage{
load_financial_indicators()

load_financial_indicators_col_explanations()
}
\value{
data.frame containg the requested data
}
\description{
\lifecycle{defunct}

Load financial indicators for specific spatial units (NUTS2 and NUTS3) for years 1995 to 2018.

Use package \code{czso} and specifically function \code{czso::czso_get_table(dataset_id = "050101")}
to obtain the data and \code{czso::czso_get_table_schema(dataset_id = "050101")} to get the columns
description.
}
\section{Functions}{
\itemize{
\item \code{load_financial_indicators_col_explanations()}: Load description for columns

}}
