% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_population_age.R
\name{load_population_age}
\alias{load_population_age}
\alias{load_population_age_col_explanations}
\title{Defunct: Load population by age}
\usage{
load_population_age(year = NA, area_type = NA)

load_population_age_col_explanations()
}
\arguments{
\item{year}{for which the data should be obtained. Default value is \code{NA}, which means
all the years. Values from range (including both limits) 2010 - 2018 are accepted.}

\item{area_type}{type of are for which the data should be obtained. Default value is \code{NA},
which means all areas. Accepted values are in Description.}
}
\value{
data.frame containg the requested data
}
\description{
\lifecycle{defunct}

Load population by sex and five-year age categories for years 2010 to 2018.

Use package \code{czso} and specifically function \code{czso::czso_get_table(dataset_id = "130142")}
to obtain the data and \code{czso::czso_get_table_schema(dataset_id = "130142")} to get the columns
description.
}
\details{
The \code{area_type} can have values from following set:
\enumerate{
\item \code{"okresy"}
\item \code{"kraje"}
\item \code{"republika"}
}
}
\section{Functions}{
\itemize{
\item \code{load_population_age()}: Load the data

\item \code{load_population_age_col_explanations()}: Load description for columns

}}
\section{Information about dataset}{

More detailed information about data can be found at the provider's website
\url{https://www.czso.cz/csu/czso/obyvatelstvo-k-3112-podle-pohlavi-v-obcich}.
}

