% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_population_settlements.R
\name{load_population_settlements}
\alias{load_population_settlements}
\alias{load_population_settlements_col_explanations}
\title{Defunct: Load populations from settlements}
\usage{
load_population_settlements(year = NA)

load_population_settlements_col_explanations()
}
\arguments{
\item{year}{year for which the data should be obtained. Default value is \code{NA}, which means
all the years. Values from range (including both limits) 2000 - 2018 are accepted.}
}
\value{
data.frame containg the requested data
}
\description{
\lifecycle{defunct}

Load population by sex for settlements for each years from 2000 to 2018.

Use package \code{czso} and specifically function \code{czso::czso_get_table(dataset_id = "130149")}
to obtain the data and \code{czso::czso_get_table_schema(dataset_id = "130149")} to get the columns
description.
}
\section{Functions}{
\itemize{
\item \code{load_population_settlements}: Load the data

\item \code{load_population_settlements_col_explanations}: Load description for columns
}}

\section{Information about dataset}{

More detailed information about data can be found at the provider's website
\url{https://www.czso.cz/csu/czso/obyvatelstvo-podle-petiletych-vekovych-skupin-a-pohlavi-v-krajich-a-okresech}.
}

