% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_RUIAN.R
\name{load_RUIAN_settlement}
\alias{load_RUIAN_settlement}
\title{Extract data from RUIAN}
\usage{
load_RUIAN_settlement(id, layer = "obec", WGS84 = FALSE)
}
\arguments{
\item{id}{id of settlement as character}

\item{layer}{identification of data to extract as character, see details.
Default value is \code{"obec"}}

\item{WGS84}{convert data to WGS-84 coordinate system? Default \code{FALSE}.}
}
\value{
\code{data.frame} with spatial objects (\code{\link[sf]{sf}}) of the specified layer
}
\description{
Extract specific layer, in form of spatial data, from RUIAN for given settlement in
Czech Republic. Checks are performed to find out if the provided \code{id} is valid for some
settlement in Czech Republic.
}
\details{
In case of adress places (using ADRM_B or adresni mista as layer) are checked, the csv file with
more attributes is also downloaded and linked to the spatial data layer.

The \code{layer} can have values from following set, the value in brackets is alias to full layer
name:
\enumerate{
\item \code{"ADRM_B"} (\code{"adresni mista"})
\item \code{"CO_B"} (\code{"casti obce"})
\item \code{"KATUZE_P"} (\code{"katastralni uzemi"})
\item \code{"OBEC_P"} (\code{"obec"})
\item \code{"SO_B"} (\code{"stavebni objekty"})
\item \code{"UL_L"} (\code{"ulice"})
\item \code{"VO_P"} (\code{"volebni okrsky"})
\item \code{"ZSJ_P"} (\code{"zakladni sidelni jednotky"})
\item \code{"MOMC_P"}
\item \code{"MOP_P"}
\item \code{"SOP_P"}
}
So the codes \code{layer = "CO_B"} and \code{layer = "casti obce"} are equal.

The values of \code{id} follow general pattern of six number with first number being 5.
}
\section{Information about dataset}{

More detailed information about data can be found at the provider's website
\url{http://atom.cuzk.cz/}.
}

\examples{
\dontrun{
    adresy_vyskov <- load_RUIAN_settlement("592889", layer = "adresni mista")
}
}
