% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_RUIAN_state.R
\name{load_RUIAN_state}
\alias{load_RUIAN_state}
\title{Extract data from RUIAN for whole Czech Republic}
\usage{
load_RUIAN_state(layer = "stat", WGS84 = FALSE)
}
\arguments{
\item{layer}{identification of data to extract as character, see details.
Default value is \code{"stĂˇt"}}

\item{WGS84}{convert data to WGS-84 coordinate system? Default \code{FALSE}.}
}
\value{
\code{data.frame} with spatail objects (\code{\link[sf]{sf}}) of the specified layer
}
\description{
Extract specific layer from RUIAN for whole Czech Republic.
}
\details{
The \code{layer} can have values from following set, the value in brackets is alias to full layer
name:
\enumerate{
\item KATUZE_P (katastralni uzemi)
\item OBCE_P (obce)
\item OKRESY_P (okresy)
\item ORP_P (orp)
\item POU_P (pou)
\item PRARES_P
\item REGION_P (regiony)
\item STATY_P (stat)
\item STU_P (stavebni úrady)
\item VO_P (volebni okrsky)
\item VUSC_P.shp (kraje)
}

So the codes \code{layer = "OKRESY_P"} and \code{layer = "okresy"} are equal.
}
\examples{
obce_CR <- load_RUIAN_state(layer = "obce")
}
