% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_orp.R
\docType{data}
\name{orp}
\alias{orp}
\title{data.frame of all settlements of type III (orp) in Czech Republic}
\format{A data frame with 206 rows and 4 variables:
\describe{
\item{kod}{id of the settlements of type III (orp)}
\item{nazev}{name of the settlements of type III (orp)}
\item{vusc_kod}{}
\item{nuts3_kod}{}
}}
\source{
\url{http://services.cuzk.cz/shp/stat/epsg-5514/1.zip}
}
\usage{
orp
}
\description{
A dataset containing the names and other attributes of all 206
settlements of type III (orp) in Czech Republic. The codes (every column with string kod in
name) are treated as character strings even though that some of them are numbers. These codes,
however, serve only as IDs.
Colums with suffix _kod are various levels of self-government units in Czech Republic.
}
\keyword{datasets}
