% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Data200.R, R/load_Data50.R
\name{generate_Data200_citation}
\alias{generate_Data200_citation}
\alias{generate_Data50_citation}
\title{Generate attribution for dataset Data50 od Data200}
\usage{
generate_Data200_citation()

generate_Data50_citation()
}
\value{
\code{character} with citation.
}
\description{
Create citation string as per terms of use (\url{https://geoportal.cuzk.cz/Dokumenty/Podminky_EN.pdf}).
}
\section{Functions}{
\itemize{
\item \code{generate_Data200_citation}: Generate citation for Data200 datasource.

\item \code{generate_Data50_citation}: Generate citation for Data50 datasource.
}}

\examples{
generate_Data50_citation()
generate_Data200_citation()

}
