% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_average_salary.R
\name{load_average_salary}
\alias{load_average_salary}
\alias{load_average_salary_col_explanations}
\title{Defunct: Load average salary for specific spatial units}
\usage{
load_average_salary()

load_average_salary_col_explanations()
}
\value{
data.frame containg the requested data
}
\description{
\lifecycle{defunct}

Load average salary for specific spatial units for years 2011 to 2017.

Use package \code{czso} and specifically function \code{czso::czso_get_table(dataset_id = "110080")}
to obtain the data and \code{czso::czso_get_table_schema(dataset_id = "110080")} to get the columns
description.
}
\section{Functions}{
\itemize{
\item \code{load_average_salary_col_explanations()}: Load description for columns

}}
