% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cadastre.R
\name{load_cadastral_territory}
\alias{load_cadastral_territory}
\title{Extract data from Cadastral map}
\usage{
load_cadastral_territory(id, layer = "katastralni uzemi",
  WGS84 = FALSE)
}
\arguments{
\item{id}{id of cadastral territory as character}

\item{layer}{identification of data to extract as character, see details.
Default value is \code{"katastralni ĂşzemĂ­"}}

\item{WGS84}{convert data to WGS-84 coordinate system? Default \code{FALSE}.}
}
\value{
\code{data.frame} with spatail objects (\code{\link[sf]{sf}}) of the specified layer
}
\description{
Extract specific layer from cadastral map for given cadastral territory in Czech Republic.
Checks are performed to find out if the provided \code{id} is valid for some cadastral territory
in Czech Republic.
}
\details{
The \code{layer} can have values from following set, the value in brackets is alias to full layer
name:
\enumerate{
\item BODOVE_POLE_B
\item BODOVE_POLE_T
\item BUDOVY_B (budovy body)
\item BUDOVY_DEF
\item BUDOVY_P (budovy)
\item DALSI_PRVKY_MAPY_B
\item DALSI_PRVKY_MAPY_L
\item DALSI_PRVKY_MAPY_T
\item HRANICE_PARCEL_L (hranice parcel)
\item KATASTRALNI_UZEMI_DEF
\item KATASTRALNI_UZEMI_L
\item KATASTRALNI_UZEMI_P (katastralni uzemi)
\item PARCELY_KN_B
\item PARCELY_KN_DEF
\item PARCELY_KN_L
\item PARCELY_KN_P (parcely)
\item PRVKY_ORIENT_MAPY_B
\item PRVKY_ORIENT_MAPY_L
\item PRVKY_ORIENT_MAPY_T
\item VB_P
}
So the codes \code{layer = "BUDOVY_B"} and \code{layer = "budovy body"} are equal.

The values of \code{id} follow general pattern of six number with first number being 6,7 or 9.
}
\examples{
parcely_vyskov <- load_cadastral_territory("788571", layer = "parcely")
}
