% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_financial_indicators.R
\name{load_financial_indicators}
\alias{load_financial_indicators}
\alias{load_financial_indicators_col_explanations}
\title{Load financial indicators for specific spatial units}
\usage{
load_financial_indicators()

load_financial_indicators_col_explanations()
}
\value{
data.frame containg the requested data
}
\description{
Load financial indicators for specific spatial units (NUTS2 and NUTS3) for years 1995 to 2018.
}
\section{Functions}{
\itemize{
\item \code{load_financial_indicators_col_explanations}: Load description for columns
}}

\examples{
\dontrun{
    average_salary <- load_average_salary()
}
}
