% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_population_settlements.R
\name{load_population_settlements}
\alias{load_population_settlements}
\alias{load_population_settlements_col_explanations}
\title{Load populations from settlements}
\usage{
load_population_settlements(year = NA)

load_population_settlements_col_explanations()
}
\arguments{
\item{year}{year for which the data should be obtained. Default value is \code{NA}, which means
all the years. Values from range (including both limits) 2000 - 2017 are accepted.}
}
\value{
data.frame containg the requested data
}
\description{
Load population by sex for settlements for each years from 2000 to 2017.
}
\section{Functions}{
\itemize{
\item \code{load_population_settlements}: Load the data

\item \code{load_population_settlements_col_explanations}: Load description for columns
}}

\section{Information about dataset}{

More detailed information about data can be found at the provider's website
\url{https://www.czso.cz/csu/czso/obyvatelstvo-podle-petiletych-vekovych-skupin-a-pohlavi-v-krajich-a-okresech}.
}

\examples{
# population for all years
population <- load_population_settlements()

# population for year 2015
population_2015 <- load_population_settlements(year = 2015)
}
