% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_RUIAN.R
\name{load_RUIAN_settlement}
\alias{load_RUIAN_settlement}
\title{Extract data from RUIAN}
\usage{
load_RUIAN_settlement(id, layer = "obec", WGS84 = FALSE)
}
\arguments{
\item{id}{id of settlement as character}

\item{layer}{identification of data to extract as character, see details.
Default value is \code{"obec"}}

\item{WGS84}{convert data to WGS-84 coordinate system? Default \code{FALSE}.}
}
\value{
\code{data.frame} with spatail objects (\code{\link[sf]{sf}}) of the specified layer
}
\description{
Extract specific layer from RUIAN for given settlement in Czech Republic. Checks are performed to
find out if the provided \code{id} is valid for some settlement in Czech Republic.
}
\details{
The \code{layer} can have values from following set, the value in brackets is alias to full layer
name:
\enumerate{
\item ADRM_B (adresni mista)
\item CO_B (casti obce)
\item KU_P (katastralni uzemi)
\item OBEC_P (obec)
\item SO_B (stavebni objekty)
\item UL_L (ulice)
\item VO_P (volebni okrsky)
\item ZSJ_P (zakladni sidelni jednotky)
}
So the codes \code{layer = "CO_B"} and \code{layer = "casti obce"} are equal.

The values of \code{id} follow general pattern of six number with first number being 5.
}
\examples{
adresy_vyskov <- load_RUIAN_settlement("592889", layer = "adresni mista")
}
