% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_RUIAN_state.R
\name{load_RUIAN_state}
\alias{load_RUIAN_state}
\title{Extract data from RUIAN for whole Czech Republic}
\usage{
load_RUIAN_state(layer = "stat", WGS84 = FALSE)
}
\arguments{
\item{layer}{identification of data to extract as character, see details.
Default value is \code{"stĂˇt"}}

\item{WGS84}{convert data to WGS-84 coordinate system? Default \code{FALSE}.}
}
\value{
\code{data.frame} with spatail objects (\code{\link[sf]{sf}}) of the specified layer
}
\description{
Extract specific layer, in form of spatial data, from RUIAN for whole Czech Republic.
The minor issue with these data is the size, the datasets that needs to be downloaded is roughly
190 MB.
}
\details{
The \code{layer} can have values from following set, the value in brackets is alias to full layer
name:
\enumerate{
\item \code{"KATUZE_P"} (\code{"katastralni uzemi"})
\item \code{"OBCE_P"} (\code{"obce"})
\item \code{"OKRESY_P"} (\code{"okresy"})
\item \code{"ORP_P"} (\code{"orp"})
\item \code{"POU_P"} (\code{"pou"})
\item \code{"PRARES_P"}
\item \code{"REGION_P"} (\code{"regiony"})
\item \code{"STATY_P"} (\code{"stat"})
\item \code{"STU_P"} (\code{"stavebni Ăşrady"})
\item \code{"VO_P"} (\code{"volebni okrsky"})
\item \code{"VUSC_P"} (\code{"kraje"})
}

So the codes \code{layer = "OKRESY_P"} and \code{layer = "okresy"} are equal.
}
\section{Information about dataset}{

More detailed information about data can be found at the provider website
\url{http://atom.cuzk.cz/}.
}

\examples{
obce_CR <- load_RUIAN_state(layer = "obce")
}
