% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_SLDB.R
\name{load_SLDB_2011}
\alias{load_SLDB_2011}
\alias{load_SLDB_2011_col_explanations}
\title{Get information from Czech census in 2011}
\usage{
load_SLDB_2011(type = "obyvatelstvo", load_names = TRUE)

load_SLDB_2011_col_explanations(type = "obyvatelstvo")
}
\arguments{
\item{type}{type of requested information as character. Default value is "obyvatelstvo".
See details for more.}

\item{load_names}{boolean value if the column names should be loaded from external source.
Default TRUE as the column names do no make any sense otherwise.}
}
\value{
data.frame containg the requested data
}
\description{
Data from Czech census in year 2011 by four main topics. The data are provided at various
aggregation levels that can be filtered.
}
\details{
Types of data that can be downloaded, and used as \code{type} parameter in the function call:
\enumerate{
\item \code{"obyvatelstvo"}
\item \code{"domy-byty"}
\item \code{"domacnosti"}
\item \code{"vyjizdka"}
}
}
\section{Functions}{
\itemize{
\item \code{load_SLDB_2011}: Load the data

\item \code{load_SLDB_2011_col_explanations}: Get names of columns for SLDB of specific type as data.frame
}}

\section{Information about dataset}{

More detailed information about data can be found at the provider's website
\url{https://www.czso.cz/csu/sldb}.
}

\examples{
sldb <- load_SLDB_2011(type = "obyvatelstvo")
}
